+++ src/lib/IExecDealModule.d.ts
@@ -207,6 +207,8 @@ export default class IExecDealModule extends IExecModule {
    */
   computeTaskId(dealid: Dealid, taskIdx: BNish): Promise<Taskid>;
   /**
+   * **SIGNER REQUIRED**
+   *
    * claim all the failed task from a deal
    *
    * depending the number and the status of task to claim, this may involve several transactions in order to fit in the blockchain gasLimit per block. (for example a 10_000_000 gas block size allows to claim 180 initialized task or 40 non-initialized tasks in one block)